Rename Import handler to assets.php
