+++ include/model/assets.php
